Move s3Uploader to lib/s3Uploader.js
